master - cart start

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,14 +3,23 @@ const jwt = require('jsonwebtoken')
 const {randomBytes} = require('crypto')
 const {promisify} = require('util')
 const {transport, makeANiceEmail} = require('../mail')
+const {hasPermission} = require('../utils')
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
-    // TODO: Check if they are logged in
+    if (!ctx.request.userId){
+      throw new Error("You have to be logged in to do that!")
+    }
     
     const item = await ctx.db.mutation.createItem(
       {
         data: {
+          //This is how we create relationships between user and item
+          user:{
+            connect:{
+              id: ctx.request.userId
+            }
+          },
           ...args,
         },
       },
@@ -34,8 +43,16 @@ const Mutations = {
   },
   async deleteItem(parent, args, ctx, info){
     const where = {id: args.id}
-    const item = await ctx.db.query.item({where}, '{id, title}')
-    //TODO check permissions
+    const item = await ctx.db.query.item({where}, '{id title user {id}}')
+    //Check permissions
+    
+    const ownsItem = item.user.id === ctx.request.userId
+    const hasPermissions = ctx.request.user.permissions.some(p => ['ADMIN','ITEMDELETE'].includes(p))
+    
+    if (!ownsItem && !hasPermissions){
+      throw new Error('You don\'t have permission to do that!')
+    }
+    
     return ctx.db.mutation.deleteItem({where}, info)
   },
   async signup(parent, args, ctx, info){
@@ -136,6 +153,21 @@ const Mutations = {
       maxAge: 1000 * 60 * 60 * 24 * 365,
     })
     return updatedUser
+  },
+  async updatePermissions(parent, args, ctx, info){
+    if(!ctx.request.userId){
+      throw new Error('You have to be logged in to continue.')
+    }
+    const currentUser = await ctx.db.query.user({
+      where: {id: ctx.request.userId}
+    }, info)
+    hasPermission(currentUser, ['ADMIN','PERMISSIONUPDATE'])
+    return ctx.db.mutation.updateUser({
+      data:{
+        permissions: {set: args.permissions}
+      },
+      where:{id: args.userId}
+    },info)
   }
 };
